+++ Tests/Fixtures/TestProject/App_supportedDestinations/Sources/iOS/ContentView.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundColor(.accentColor)
+            Text("Hello, world! on iOS")
+        }
+        .padding()
+    }
+}